+++ CHANGELOG.md
@@ -10,6 +10,11 @@ speaking, considered to be the player's experience.
 
 ## [Unreleased]
 
+### Changed
+
+- Changed default keybind for train soldier to `F` from `W`, to make it more
+  consistent with the other fort-related keybinds (build fort).
+
 
 ## [3.2.3] — 2025-08-24